Asignacion de permisos a usuarios

+++ src/app/Modules/protected/components/socios/socios.component.ts
@@ -8,6 +8,7 @@ import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
 import Swal from 'sweetalert2';
 import { DataResponse } from 'src/app/Interfaces/data';
+import { TokenService } from 'src/app/Services/token.service';
 
 
 @Component({
@@ -41,7 +42,7 @@ export class SociosComponent implements OnInit {
   @ViewChild(MatPaginator) paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
 
-  constructor(private partnerS: PartnerService, private router: Router) {
+  constructor(private partnerS: PartnerService, private router: Router, private tokenS: TokenService) {
     // Create 100 users
     //const users = Array.from({length: 100}, (_, k) => createNewUser(k + 1));
     // Assign the data to the data source for the table to render
@@ -69,35 +70,46 @@ export class SociosComponent implements OnInit {
   
 
   removeData(id: number) {
-    this.partnerS.getPartner(id).subscribe((res) => {
-      this.responsePartner = res;
+    const role = this.tokenS.getAutorities();
+    role.forEach( rol =>{
+      if(rol === 'ROL_ADMINISTRATOR'){
 
-      Swal.fire({
-        title: `Desea eliminar a socio ${this.responsePartner.responsePartner.name}`,
-        icon: 'warning',
-        showCancelButton: true,
-        cancelButtonText: 'Cancelar',
-        confirmButtonText: 'Eliminar',
-        confirmButtonColor: '#3085d6',
-        cancelButtonColor: '#d33',
-      }).then((result) => {
-        if (result.value) {
-          this.partnerS.deletePartner(id).subscribe({
-            next: () => {
-              this.action = 'eliminado';
-
-              this.toast(
-                this.responsePartner.responsePartner.name,
-                this.action
-              );
-              this.getAllPartners();
-              
-            },
+        this.partnerS.getPartner(id).subscribe((res) => {
+          this.responsePartner = res;
+    
+          Swal.fire({
+            title: `Desea eliminar a socio ${this.responsePartner.responsePartner.name}`,
+            icon: 'warning',
+            showCancelButton: true,
+            cancelButtonText: 'Cancelar',
+            confirmButtonText: 'Eliminar',
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+          }).then((result) => {
+            if (result.value) {
+              this.partnerS.deletePartner(id).subscribe({
+                next: () => {
+                  this.action = 'eliminado';
+    
+                  this.toast(
+                    this.responsePartner.responsePartner.name,
+                    this.action
+                  );
+                  this.getAllPartners();
+                  
+                },
+              });
+            }
           });
-        }
-      });
-    });
+        });
+    
 
+      }else{
+
+        this.toastE()
+      }
+    })
+    
     
   }
 
@@ -116,7 +128,7 @@ export class SociosComponent implements OnInit {
       toast: true,
       position: 'center',
       showConfirmButton: false,
-      timer: 3000,
+      timer: 2000,
       timerProgressBar: true,
       didOpen: (toast) => {
         toast.addEventListener('mouseenter', Swal.stopTimer);
@@ -127,6 +139,25 @@ export class SociosComponent implements OnInit {
     Toast.fire({
       icon: 'success',
       title: `Socio ${name} ${action} exitosamente¡¡¡¡¡¡ `,
+    });
+  }
+
+  toastE(): void {
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'center',
+      showConfirmButton: false,
+      timer: 3000,
+      timerProgressBar: true,
+      didOpen: (toast) => {
+        toast.addEventListener('mouseenter', Swal.stopTimer);
+        toast.addEventListener('mouseleave', Swal.resumeTimer);
+      },
+    });
+
+    Toast.fire({
+      icon: 'error',
+      title: `Permiso Denegado¡¡¡¡¡¡ `,
     });
   }
 }
